+++ src/card/sets/y2021/mod.rs
@@ -1,6 +1,7 @@
 //! Card sets first released in 2021.
 
 pub(super) mod adventures_in_the_forgotten_realms;
+pub(super) mod commander_2021;
 pub(super) mod innistrad_crimson_vow_commander;
 pub(super) mod innistrad_midnight_hunt;
 pub(super) mod kaldheim;