+++ src/game/tests.rs
@@ -824,6 +824,7 @@ mod vintage_cube_infect;
 mod vintage_cube_jace_the_mind_sculptor;
 mod vintage_cube_jacked_rabbit;
 mod vintage_cube_jitte;
+mod vintage_cube_laelia;
 mod vintage_cube_lands;
 mod vintage_cube_lavaspur_boots;
 mod vintage_cube_ledger_shredder;